How to not revalidate the response?
I had a question here, rather than any pull request.

Once we past max-age since the response was generated from server we dont want to revalidate with conditional-get from the backend. Instead we want to just send get call to the backend. Is there a way to achieve this (like by setting some kind of attribute in client config)?